extract IssuesReplayCountProvider to its own file
