+++ static/app/components/replays/replayCountContext.tsx
@@ -1,11 +1,6 @@
-import {createContext, ReactNode, useMemo} from 'react';
-import first from 'lodash/first';
+import {createContext} from 'react';
 
-import useReplaysCount from 'sentry/components/replays/useReplaysCount';
-import GroupStore from 'sentry/stores/groupStore';
-import type {Group} from 'sentry/types';
-import projectSupportsReplay from 'sentry/utils/replays/projectSupportsReplay';
-import useOrganization from 'sentry/utils/useOrganization';
+import type useReplaysCount from 'sentry/components/replays/useReplaysCount';
 
 /**
  * To set things up:
@@ -26,42 +21,3 @@ import useOrganization from 'sentry/utils/useOrganization';
 const ReplayCountContext = createContext<ReturnType<typeof useReplaysCount>>({});
 
 export default ReplayCountContext;
-
-/**
- * Component to make it easier to query for replay counts against a list of groups
- * that exist across many projects.
- */
-export function ReplayCountProvider({
-  children,
-  groupIds,
-}: {
-  children: ReactNode;
-  groupIds: string[];
-}) {
-  const organization = useOrganization();
-
-  // Only ask for the groupIds where the project supports replay.
-  // For projects that don't support replay the count will always be zero.
-  const groups = useMemo(
-    () =>
-      groupIds
-        .map(id => GroupStore.get(id) as Group)
-        .filter(Boolean)
-        .filter(group => projectSupportsReplay(group.project)),
-    [groupIds]
-  );
-
-  // Any project that supports replay will do here.
-  // Project is used to signal if we should/should not do the query at all.
-  const project = first(groups)?.project;
-
-  const counts = useReplaysCount({
-    groupIds,
-    organization,
-    project,
-  });
-
-  return (
-    <ReplayCountContext.Provider value={counts}>{children}</ReplayCountContext.Provider>
-  );
-}
