+++ static/app/views/alerts/rules/issue/previewTable.tsx
@@ -6,7 +6,7 @@ import GroupListHeader from 'sentry/components/issues/groupListHeader';
 import LoadingIndicator from 'sentry/components/loadingIndicator';
 import Pagination, {CursorHandler} from 'sentry/components/pagination';
 import {Panel, PanelBody} from 'sentry/components/panels';
-import {ReplayCountProvider} from 'sentry/components/replays/replayCountContext';
+import IssuesReplayCountProvider from 'sentry/components/replays/issuesReplayCountProvider';
 import StreamGroup from 'sentry/components/stream/group';
 import {t, tct} from 'sentry/locale';
 import GroupStore from 'sentry/stores/groupStore';
@@ -95,13 +95,13 @@ const PreviewTable = ({
   };
 
   return (
-    <ReplayCountProvider groupIds={previewGroups || []}>
+    <IssuesReplayCountProvider groupIds={previewGroups || []}>
       <Panel>
         <GroupListHeader withChart={false} />
         <PanelBody>{renderBody()}</PanelBody>
       </Panel>
       {renderPagination()}
-    </ReplayCountProvider>
+    </IssuesReplayCountProvider>
   );
 };
